Remove manual check for permissions listed in manifest
This is not necessary because the permissions defined in the
sdk are merged into the app's manifest. This change can cause an
issue when someone manually removes them with
tools:node="remove", but those cases are fairly uncommon, it might
be better to warn against that in the README so api specific lists
in the code are not necessary.